Complete first version of install.sh
All admin skel is stored in linuxclientsetup/admin-skel, so that needs to be
copied to ~administrator before remasters.

install.sh can now perform remasters, correctly cleaning up after itself once
done. It automatically adds and removes the karoshi-setup autostart links, as
well as getting the current version and website listed in README.md. Processor
architecture is determined using the output of `uname -i`, but x86_64 is
converted into amd64 for consistency with ISO naming standards. Not sure if
1386/i486/i586/i686/x86 needs to be converted also.