Recode package installation logic, make messages consistent in install.sh
apt-mark is now used to ensure that packages to be removed are not installed
as recommends by the install section.

Also, all messages output by the script go to stderr, so they can appear on
the console even if stdout is redirected by the user.

Error codes also get printed should a command fail.