Change karoshi-update to run full-time, fix CUPS issue
karoshi-update is now a proper job that runs all the time the system is up.
It also only runs when the system has updated successfully, allowing other
jobs to use the stanza 'started karoshi-update' to detect if the system is
online and connected to the network, and has updated.

Included in this update is a fix for the persistent CUPS issue where CUPS
karoshi-cups will not run if CUPS doesn't start on time/starts too quickly.
That was caused by Upstart being silly - now the dependencies are much
simpler, with karoshi-cups starting when cups and karoshi-update are both
started.