Generate git system configuration for proxy settings.
During client configuration the system git configuration (/etc/gitconfig) will be modified (and created if necessary) so that the proxy settings will be used with git. This will not be run if a proxy server is not present. The script will check whether this action has been performed before to prevent duplicate configurations.