Client modified to use GVFS for mounting network shares
Mounting network shares with pam_mount uses mount-cifs. Currently mount-cifs has
a problem with copying folders with certain files such as .txt files in them.
This causes the copy process to either hang or go very slowly depending on the
smb version used to connect to the share. The only version that works is 1.0 which
does not seem to support acl permissions on folders.

As a workaround until this issue is resolved the client has been configured to use GVFS
for connecting to network shares which does not use mount.cifs for connecting to the share.
The downside to this is that GVFS cannot currently do domain DFS shares.